# backend/backend/docs/swagger/messenger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Suda Messenger: authentication, chats, messaging, users, search.",
        "title": "Suda Messenger Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/users/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "List my friends",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size (1..100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.FriendInfo"
                            }
                        }
                    }
                }
            }
        },
        "/users/friends/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "List pending friend requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "incoming | outgoing (default: both)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size (1..100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.FriendRequestInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a PENDING request to target user. Errors if already friends / pending / self / blocked.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Send friend request",
                "parameters": [
                    {
                        "description": "Target user id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.SendRequestReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.FriendRequest"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR | SELF_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "BLOCKED",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "ALREADY_FRIENDS | ALREADY_PENDING",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/friends/requests/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requester withdraws their own PENDING request.",
                "tags": [
                    "Friends"
                ],
                "summary": "Cancel sent friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    }
                }
            }
        },
        "/users/friends/requests/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Target user accepts. After this both users are FRIENDS.",
                "tags": [
                    "Friends"
                ],
                "summary": "Accept friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "403": {
                        "description": "FORBIDDEN",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "INVALID_TRANSITION",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/friends/requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Reject friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    }
                }
            }
        },
        "/users/friends/{user_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Remove friend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Friend user UUID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/users/friends/{user_id}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns NONE / PENDING_SENT / PENDING_RECEIVED / FRIENDS / BLOCKED.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Get relation status with a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Other user UUID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.RelationStatusResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.FriendInfo": {
            "type": "object",
            "properties": {
                "avatar_media_id": {
                    "type": "string"
                },
                "became_at": {
                    "description": "ISO time when request was accepted",
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "is_online": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.FriendRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "requester_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.FriendRequestInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "direction": {
                    "description": "\"incoming\" | \"outgoing\"",
                    "type": "string"
                },
                "other_avatar_media_id": {
                    "type": "string"
                },
                "other_display_name": {
                    "type": "string"
                },
                "other_user_id": {
                    "type": "string"
                },
                "other_username": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "description": "PENDING / ACCEPTED / REJECTED / CANCELLED",
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.RelationStatusResponse": {
            "type": "object",
            "properties": {
                "relation": {
                    "description": "see RelationXxx constants",
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_friends.SendRequestReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}